EV-49 move EventWrapper state props to containers, change name to camelCase, add btn title prop for NoResult
