+++ src/components/shared/NoResult/NoResult.jsx
@@ -1,12 +1,9 @@
-// import { createBrowserHistory } from "history";
 import React from "react";
 import { useHistory } from "react-router-dom";
 import PropTypes from "prop-types";
 import "./NoResult.scss";
 
-const NoResult = ({ title, subTitle, photo }) => {
-  // const history = createBrowserHistory({ forceRefresh: true });
-  // onClick={() => history.push("/landing")}
+const NoResult = ({ title, subTitle, photo, btnTitle }) => {
   const history = useHistory();
   return (
     <>
@@ -20,7 +17,7 @@ const NoResult = ({ title, subTitle, photo }) => {
         onClick={() => history.push("/landing")}
         type="button"
       >
-        Back
+        {btnTitle}
       </button>
     </>
   );
@@ -30,12 +27,14 @@ NoResult.propTypes = {
   title: PropTypes.string,
   subTitle: PropTypes.string,
   photo: PropTypes.string,
+  btnTitle: PropTypes.string,
 };
 
 NoResult.defaultProps = {
   title: "",
   subTitle: "",
   photo: "",
+  btnTitle: "",
 };
 
 export default NoResult;
